backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pymongo==3.12.0
python-dotenv==0.19.0
flask-jwt-extended==4.3.1
bcrypt==3.2.0
python-jose==3.3.0
catboost==1.2.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
